requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
colorama==0.4.4
cryptography==3.4.7
DateTime==4.3
defusedxml==0.7.1
Deprecated==1.2.12
Django==3.2.5
django-allauth==0.45.0
django-cors-headers==3.7.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
et-xmlfile==1.1.0
exchange-calendars==3.4
fake-useragent==0.1.11
finance-datareader==0.9.31
html5lib==1.1
idna==3.2
install==1.3.4
korean-lunar-calendar==0.2.1
lxml==4.6.3
numpy==1.21.1
oauthlib==3.1.1
openpyxl==3.0.9
pandas==1.3.4
pycparser==2.20
PyJWT==1.7.1
pykrx==1.0.18
pyluach==1.3.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-file==1.5.1
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
toolz==0.11.1
tqdm==4.61.2
typing-extensions==3.10.0.0
urllib3==1.26.6
webencodings==0.5.1
wrapt==1.12.1
xlrd==2.0.1
zope.interface==5.4.0
